feat: refactor service provider code
