+++ src/provider/context.ts
@@ -1,4 +1,4 @@
 import { createContext } from 'react';
-import { IServices } from './service';
+import { IServices } from './types';
 
 export const ServiceContext = createContext<IServices | null>(null);
