+++ src/provider/index.ts
@@ -1,3 +1,4 @@
 export { ServiceProvider } from './ServiceProvider';
 export { useService } from './useService';
-export type { IServices } from './service';
+export { type IServices } from './types';
+export { services } from './services';
